+++ app/shared/template.js
@@ -912,7 +912,6 @@ define(['shared/icon'], (ICON) => ({
             </form>
         </div>
 
-        <div class="download-app"></div>
         <!-- Version -->
         <div class="xm-current-version"></div>
         <!-- learn more -->